+++ flutter-app/app/android/app/src/main/kotlin/com/example/tictactoe/tic_tac_toe_app/MainActivity.kt
@@ -1,5 +1,0 @@
-package com.example.tictactoe.tic_tac_toe_app
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity : FlutterActivity()